NI-340 - add ComponentViewModel
