+++ Sources/RoktUXHelper/UI/Components/ViewModel/LayoutSchemaViewModel.swift
@@ -11,71 +11,6 @@
 
 import Foundation
 
-protocol ComponentViewModel {}
-
-@available(iOS 15, *)
-extension OverlayViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension BottomSheetViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension RowViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension ColumnViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension ZStackViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension WhenViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension OneByOneViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension CarouselViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension GroupedDistributionViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension RichTextViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension BasicTextViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension CreativeResponseViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension StaticImageViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension DataImageViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension ProgressIndicatorViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension CloseButtonViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension StaticLinkViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension ProgressControlViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension ToggleButtonViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension CatalogStackedCollectionViewModel: ComponentViewModel {}
-
-@available(iOS 15, *)
-extension CatalogResponseButtonViewModel: ComponentViewModel {}
-
 @available(iOS 15, *)
 enum LayoutSchemaViewModel: Hashable {
     // top-level
@@ -160,10 +95,6 @@ extension LayoutSchemaViewModel {
         case .empty:
             nil
         }
-    }
-
-    var isDomainMappableParent: Bool {
-        componentViewModel is DomainMappableParent
     }
 }
